Test to make json rounding more efficient

+++ MTA10/mods/shared_logic/lua/CLuaArgument.cpp
@@ -836,7 +836,7 @@ json_object * CLuaArgument::WriteToJSONObject ( bool bSerialize, CFastHashMap < 
             }
             else
             {
-                return json_object_new_double( static_cast < float > ( GetNumber() ) );
+                return json_object_new_double( GetNumber() );
             }
             break;
         }
@@ -926,7 +926,7 @@ char * CLuaArgument::WriteToString ( char * szBuffer, int length )
             }
             else
             {
-                snprintf ( szBuffer, length, "%f", static_cast < float > ( GetNumber() ) );
+                snprintf ( szBuffer, length, "%f", GetNumber() );
                 return szBuffer;
             }
             break;
